refactor: common 패키지 파일 정리
