--- UOSense-Backend/src/main/java/UOSense/UOSense_Backend/common/CategoryConverter.java
@@ -1,10 +1,0 @@
-package UOSense.UOSense_Backend.common;
-
-import jakarta.persistence.Converter;
-
-@Converter(autoApply = true)
-public class CategoryConverter extends EnumBaseConverter<Category>{
-    public CategoryConverter() {
-        super(Category.class);
-    }
-}
