--- UOSense-Backend/src/main/java/UOSense/UOSense_Backend/common/SubDescriptionConverter.java
@@ -1,7 +1,0 @@
-package UOSense.UOSense_Backend.common;
-
-public class SubDescriptionConverter extends EnumBaseConverter<SubDescription>{
-    public SubDescriptionConverter() {
-        super(SubDescription.class) ;
-    }
-}
