+++ UOSense-Backend/src/main/java/UOSense/UOSense_Backend/common/Utils/RedisUtil.java
@@ -1,4 +1,4 @@
-package UOSense.UOSense_Backend.common;
+package UOSense.UOSense_Backend.common.Utils;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
