+++ UOSense-Backend/src/main/java/UOSense/UOSense_Backend/common/Utils/SearchUtils.java
@@ -1,4 +1,4 @@
-package UOSense.UOSense_Backend.common;
+package UOSense.UOSense_Backend.common.Utils;
 
 import UOSense.UOSense_Backend.dto.RestaurantListResponse;
 import UOSense.UOSense_Backend.dto.SearchPair;
