+++ UOSense-Backend/src/main/java/UOSense/UOSense_Backend/common/converter/SubDescriptionConverter.java
@@ -1,0 +1,9 @@
+package UOSense.UOSense_Backend.common.converter;
+
+import UOSense.UOSense_Backend.common.enumClass.SubDescription;
+
+public class SubDescriptionConverter extends EnumBaseConverter<SubDescription> {
+    public SubDescriptionConverter() {
+        super(SubDescription.class) ;
+    }
+}
