+++ UOSense-Backend/src/main/java/UOSense/UOSense_Backend/common/converter/TagConverter.java
@@ -1,5 +1,6 @@
-package UOSense.UOSense_Backend.common;
+package UOSense.UOSense_Backend.common.converter;
 
+import UOSense.UOSense_Backend.common.enumClass.Tag;
 import jakarta.persistence.Converter;
 
 @Converter(autoApply = true)
