+++ UOSense-Backend/src/main/java/UOSense/UOSense_Backend/common/enumClass/BaseEnum.java
@@ -1,4 +1,4 @@
-package UOSense.UOSense_Backend.common;
+package UOSense.UOSense_Backend.common.enumClass;
 
 public interface BaseEnum {
     String getValue();
