+++ UOSense-Backend/src/main/java/UOSense/UOSense_Backend/common/enumClass/Detail.java
@@ -1,7 +1,5 @@
 package UOSense.UOSense_Backend.common.enumClass;
 
-import UOSense.UOSense_Backend.common.BaseEnum;
-
 public enum Detail implements BaseEnum {
     ABUSIVE("욕설"), DEROGATORY("비하 발언"), ADVERTISEMENT("광고성 리뷰");
