+++ UOSense-Backend/src/main/java/UOSense/UOSense_Backend/common/enumClass/DoorType.java
@@ -1,4 +1,4 @@
-package UOSense.UOSense_Backend.common;
+package UOSense.UOSense_Backend.common.enumClass;
 
 public enum DoorType implements BaseEnum {
     FRONT("정문"), SIDE("쪽문"), BACK("후문"), SOUTH("남문");
