+++ UOSense-Backend/src/main/java/UOSense/UOSense_Backend/common/enumClass/Role.java
@@ -1,4 +1,4 @@
-package UOSense.UOSense_Backend.common;
+package UOSense.UOSense_Backend.common.enumClass;
 
 public enum Role implements BaseEnum {
     USER, ADMIN;
