+++ UOSense-Backend/src/main/java/UOSense/UOSense_Backend/common/enumClass/SubDescription.java
@@ -1,6 +1,6 @@
-package UOSense.UOSense_Backend.common;
+package UOSense.UOSense_Backend.common.enumClass;
 
-public enum SubDescription implements BaseEnum{
+public enum SubDescription implements BaseEnum {
     BAR("술집"), CAFE("카페"), RESTAURANT("음식점");
     private final String value;
