+++ UOSense-Backend/src/main/java/UOSense/UOSense_Backend/controller/RestaurantController.java
@@ -1,7 +1,7 @@
 package UOSense.UOSense_Backend.controller;
 
-import UOSense.UOSense_Backend.common.Category;
-import UOSense.UOSense_Backend.common.DoorType;
+import UOSense.UOSense_Backend.common.enumClass.Category;
+import UOSense.UOSense_Backend.common.enumClass.DoorType;
 import UOSense.UOSense_Backend.dto.*;
 import UOSense.UOSense_Backend.entity.Restaurant;
 import UOSense.UOSense_Backend.service.MenuService;
