+++ UOSense-Backend/src/main/java/UOSense/UOSense_Backend/controller/SearchController.java
@@ -1,6 +1,6 @@
 package UOSense.UOSense_Backend.controller;
 
-import UOSense.UOSense_Backend.common.DoorType;
+import UOSense.UOSense_Backend.common.enumClass.DoorType;
 import UOSense.UOSense_Backend.dto.RestaurantListResponse;
 import UOSense.UOSense_Backend.entity.Restaurant;
 import UOSense.UOSense_Backend.service.SearchService;
