+++ UOSense-Backend/src/main/java/UOSense/UOSense_Backend/service/RestaurantImageServiceImpl.java
@@ -1,6 +1,6 @@
 package UOSense.UOSense_Backend.service;
 
-import UOSense.UOSense_Backend.common.ImageUtils;
+import UOSense.UOSense_Backend.common.Utils.ImageUtils;
 import UOSense.UOSense_Backend.dto.ImageInfo;
 import UOSense.UOSense_Backend.dto.RestaurantImagesResponse;
 import UOSense.UOSense_Backend.entity.Restaurant;
