+++ UOSense-Backend/src/main/java/UOSense/UOSense_Backend/service/RestaurantService.java
@@ -1,7 +1,7 @@
 package UOSense.UOSense_Backend.service;
 
-import UOSense.UOSense_Backend.common.Category;
-import UOSense.UOSense_Backend.common.DoorType;
+import UOSense.UOSense_Backend.common.enumClass.Category;
+import UOSense.UOSense_Backend.common.enumClass.DoorType;
 import UOSense.UOSense_Backend.dto.*;
 import UOSense.UOSense_Backend.entity.Restaurant;
