+++ UOSense-Backend/src/main/java/UOSense/UOSense_Backend/service/RestaurantServiceImpl.java
@@ -1,7 +1,8 @@
 package UOSense.UOSense_Backend.service;
 
-import UOSense.UOSense_Backend.common.*;
-
+import UOSense.UOSense_Backend.common.Utils.SearchUtils;
+import UOSense.UOSense_Backend.common.enumClass.Category;
+import UOSense.UOSense_Backend.common.enumClass.DoorType;
 import UOSense.UOSense_Backend.dto.*;
 
 import UOSense.UOSense_Backend.entity.BusinessDay;
